fix mount units when run the first time
If $ANDROID_ROOT/units/ doesn't exist, old approach produces an RPM
which doesn't enable mount units (points like /system are not mounted
on device). Running mb2 the second time works fine (because units/
already exists)

This commit fixes the issue, and correct RPM is baked right away.

I'm not happy with this approach, but it is least worst.

Signed-off-by: Simonas Leleiva <simonas.leleiva@jollamobile.com>