[maguro][modem] UGLY* hack to work around ofono issues
As it stands now, ofono has issues reading and writing to/from the SIM.
Issues like: failing to read the IMSI, correct PIN registered as wrong PIN,
counting UP the remaining tries instead of DOWN, etc.

To allow a quick preview release for maguro, and since the developers
are busy with more important stuff: WORK AROUND IT

Users have to disable their PIN before flashing or through another device.
The script/service listens on ofonos dbus interface for changes regarding
modem power state. If the modem is powered => force the modem to online state
via a dbus call.

This is supposed to be a short-term hack. Until the developers (and I) have the
time to fix the real issue, this might do it.

*) ugly as in "hedgehog hit by a car"-ugly.

Change-Id: Ib37e2030ad4aab6df82e760e8575dfecd1c823ee