[maguro][bluetooth] Fix bluetooth enable/disable
The tuna bluetooth platform driver sends rfkill uevents when bluetooth is
supposed to be activated. Let's react to these events and re-start the
firmware download whenever bluetooth is activated.
Also, to not confuse udev, do not rfkill-unblock the bluetooth devices
(including the platform device) as it could end in an infinite loop.

Change-Id: I1c1b04c7138c20b46e3bc15a52ecb754113d7e15