Move /etc/hw-release out of droid-hal-device.
[packaging] Move /etc/hw-release out of droid-hal-device.

When doing official releases the version needs to be generated the last
in the build change to make it properly done. Thus we can not have
hw-release here but it is moved to droid-hal-version package which is
the last package that is built in the hardware adaptation repo thus it
always has the proper dependencies to other packages.

Signed-off-by: Marko Saukko <marko.saukko@jolla.com>