helpers: Add build kernel script if in tree builds are not possible
[helpers] Add build kernel script if in tree builds are not possible

Add script to build the kernel to replicate the kernel build process
like it was in aosp android 9, so that dhd builds can continue like
before.
This fixes cases where in tree builds are not possible like android 10.

The script requires some additional environment variables as listed in
line #20.

These are currently:
 - CROSS_COMPILE - path to cross compiler
 - CROSS_COMPILE_ARM32 - path to ARM32 cross compiler
 - KERNEL_DEFCONFIG - defconfig used to generate .config something like aosp_$platform_$HABUILD_DEVICE
 - KERNEL_SOURCE - path to kernel source
 - DROID_TARGET_KERNEL_ARCH - target architecture of device like droid_target_arch but different naming
 - DROID_TARGET_KERNEL_DTB - if defined dtbo.img gets generated if your device needs it

All are named to resemble either dhd names or exiting kernel build options.